% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_line.R
\name{read_line}
\alias{read_line}
\title{Read single variant from VCF file to R.}
\usage{
read_line(VCF, chr = NULL, pos = NULL, infoOnly = TRUE)
}
\arguments{
\item{VCF}{Path to target block zipped VCF file (*.vcf.gz)}

\item{chr}{The chromosome you wanted to read. Whole VCF file will be loaded by default.}

\item{pos}{The specific positions wanted to load on the target chromosome.}

\item{infoOnly}{Only load variant information. Default: TRUE. Use FALSE if you want to load genotype data.}
}
\value{
VCF file read in data frame format
}
\description{
Read single variant from block zipped VCF file to R.
}
