% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vcf.R
\name{read_vcf}
\alias{read_vcf}
\title{Read VCF file to R.}
\usage{
read_vcf(VCF, chr = NULL, from = NULL, to = NULL, infoOnly = TRUE)
}
\arguments{
\item{VCF}{Path to target block zipped VCF file (*.vcf.gz)}

\item{chr}{The chromosome you wanted to read. Whole VCF file will be loaded by default.}

\item{from}{The specific region wanted to load on the target chromosome.}

\item{to}{The specific region wanted to load on the target chromosome.}

\item{infoOnly}{Only load variant information. Default: TRUE. Use FALSE if you want to load genotype data.}
}
\value{
VCF file read in data frame format
}
\description{
Read block zipped VCF file to R.
}
